add delete itme

+++ src/app/staff/main/main.component.ts
@@ -67,26 +67,35 @@ export class MainComponent implements OnInit {
       this.alertService.error();
     }
   }
-  async removeRequest(requestId: any) {
+
+
+
+
+
+
+
+  async removeRequest(requestId) {
     this.alertService.confirm()
-      .then(async () => {
-        try {
-          let rs: any = await this.requestService.removeRequest(requestId);
-          if (rs.ok) {
-            this.alertService.removeSuccess();
-            this.getRequest(this.perPage, 0);
-          } else {
+      .then(async (result) => {
+        if (result.value) {
+          try {
+            let rs: any = await this.requestService.removeRequest(requestId);
+            if (rs.ok) {
+
+              this.alertService.success();
+              this.getRequest(this.perPage, 0);
+            } else {
+              this.alertService.error();
+            }
+
+          } catch (error) {
             this.alertService.error();
           }
-        } catch (error) {
-          this.alertService.error();
 
         }
       })
-      .catch(() => {
-        console.log('cancel');
-      });
   }
+
